\name{LBI-package}
\alias{LBI-package}
\alias{LBI}
\docType{package}
\title{Likelihood Based Inference}
\description{
It conducts likelihood based inference.
}
\details{
  Modern likelihood concept and maximum likelihood estimation are established by RA Fisher, while Likelihood Ratio Test (LRT) is established by Neyman J. Post-Fisher methods - generalized linear model, survival analysis, and mixed effects model - are all likelihood based. Inferences from the perspective of Fisherian and pure likelihoodist are suggested here.
}
\author{Kyun-Seop Bae <k@acr.kr>}

\references{
\enumerate{
\item Wilks SS. The Large-sample Distribution of the Likelihood Ratio for Testing Composite Hypotheses. Ann Math Stat. 1938;9(1):60-62.
\item Fisher RA. Statistical Methods and Scientific Inference. 3e. 1973.
\item Edwards AWF. Likelihood. 1972.
\item Ruppert D, Cressie N, Carroll RJ. A Transformation/Weighting Model for Estimating Michaelis-Menten Parameters. Cornell University Technical Report 796. 1988.
\item Lehmann EL. Fisher, Nayman, and the Creation of Classical Statistics. 2011.
\item Royall R. Statistical Evidence. 1997.
\item Pawitan Y. In All Likelihood: Statistical Modelling and Inference Using Likelihood. 2001.
\item Rohde CA. Introductory Statistical Inference with the Likelihood Function. 2014.
\item Held L, Bove DS. Likelihood and Bayesian Inference. 2020.
}
}
